Move state/theme reset fixtures to _reset_fixtures.py
Split _reset_pyvista_state, _restore_default_pyvista_state, and
_set_default_theme out of pytest_pyvista.py into a dedicated module,
and their tests into tests/test_reset_fixtures.py.
